public_backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
mariadb==1.1.8
markdownify==0.11.6
mysqlclient==2.2.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pycparser==2.21
PyMySQL==1.1.0
pyspnego==0.10.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-kerberos==0.14.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
sspilib==0.1.0
tzdata==2023.3
urllib3==2.0.6